+++ src/traits/i-history/history/index.ts
@@ -6,8 +6,8 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import Async from 'core/async';
 import symbolGenerator from 'core/symbol';
+import Async from 'core/async';
 
 import { ModsDecl, ComponentHooks } from 'core/component';
 import { InView } from 'core/component/directives/in-view';
@@ -264,7 +264,7 @@ export default class History<C extends iHistory> {
 
 		this.async.requestAnimationFrame(() => {
 			Object.assign(t.style, {
-				height: 1,
+				height: (1).px,
 				width: '100%',
 				position: 'absolute',
 				top: 0,
@@ -293,10 +293,11 @@ export default class History<C extends iHistory> {
 			InView.observe(el, {
 				threshold: this.config.titleThreshold,
 				onEnter: () => this.onPageTopVisibilityChange(true),
-				onLeave: () => this.onPageTopVisibilityChange(false)
+				onLeave: () => this.onPageTopVisibilityChange(false),
+				polling: true
 			});
 
-			this.async.worker(() => InView.stopObserve(el), label);
+			this.async.worker(() => InView.remove(el), label);
 
 		} else {
 			this.async.terminateWorker(label);